Add templates

+++ src/pipelines/prediction_pipeline.py
@@ -118,8 +118,8 @@ class CustomData:
         parental_level_of_education: str,
         lunch:                       str,
         test_preparation_course:     str,
-        reading_score:               int,
-        writing_score:               int,
+        reading_score:               float,
+        writing_score:               float,
     ):
         # Store each field as an instance attribute
         # Type annotations document exactly what the
